requirements.txt
openai
transformers
torch
scikit-learn
numpy
pandas
sentence-transformers
redis
python-dotenv
pydantic
pytest
black
flake8
jupyter
